changed type of return value back to int from int64_t due to issues thsi created with the values

+++ syscall-replayer/SystemCallTraceReplayModule.cpp
@@ -70,8 +70,8 @@ int SystemCallTraceReplayModule::errno_number() const {
   return (int)errno_number_.val();
 }
 
-int64_t SystemCallTraceReplayModule::return_value() const {
-  return (int64_t)return_value_.val();
+int SystemCallTraceReplayModule::return_value() const {
+  return (int)return_value_.val();
 }
 
 int64_t SystemCallTraceReplayModule::unique_id() const {
